+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This file is used to list changes made in each version of cookbook-openstack-orc
 * Sync conf files with Juno
 * Upgrading berkshelf from 2.0.18 to 3.1.5
 * Update mode for heat.conf from 644 to 640
+* Add cafile, memcached_servers, memcache_security_strategy, memcache_secret_key, insecure and hash_algorithms so that they are configurable.
 
 ## 9.2.0
 * python_packages database client attributes have been migrated to